+++ apps/web/.storybook/main.ts
@@ -1,4 +1,6 @@
 import type { StorybookConfig } from '@storybook/react-vite';
+import { mergeConfig } from 'vite';
+import { createSharedViteConfig } from '../vite.shared.config.js';
 
 const config: StorybookConfig = {
   stories: ['../src/**/*.stories.@(ts|tsx)'],
@@ -8,6 +10,13 @@ const config: StorybookConfig = {
     options: {},
   },
   core: { disableTelemetry: true },
+  viteFinal: (viteConfig) =>
+    Promise.resolve(
+      mergeConfig(
+        viteConfig,
+        createSharedViteConfig({ includePlugins: false }),
+      ),
+    ),
 };
 
 export default config;